% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.big.data.frame}
\alias{as.big.data.frame}
\title{Convert a \code{\link{big.matrix}} object to a \code{\link{big.data.frame}} object}
\usage{
as.big.data.frame(x, names = NULL)
}
\arguments{
\item{x}{a big.matrix object}
}
\value{
a new \code{\link{big.data.frame}} object
}
\description{
Convert a \code{\link{big.matrix}} object to a \code{\link{big.data.frame}} object
}
\author{
Miranda Sinnott-Armstrong
}

