% Generated by roxygen2 (4.0.2): do not edit by hand
\name{attach.big.data.frame}
\alias{attach.big.data.frame}
\title{Attach an existing big.data.frame from its backingfile}
\usage{
attach.big.data.frame(location)
}
\arguments{
\item{location}{the folder containing the backingfiles.}
}
\value{
a \code{\link{big.data.frame}} object
}
\description{
The expected usage is for shared-memory parallel computing
or for persistence via memory-mapped files of the (column) variables.
}

