% Generated by roxygen2 (4.0.2): do not edit by hand
\name{big.add.col}
\alias{big.add.col}
\title{Add an extra column to an existing \code{\link{big.data.frame}}}
\usage{
big.add.col(x, new.col, after, new.name, location = NULL)
}
\arguments{
\item{x}{a big.data.frame object}

\item{new.col}{either a big.char object or a big.matrix object
with a single column}

\item{after}{the position after which new.col will be appear}
}
\value{
a new \code{\link{big.data.frame}} object, with one extra column
}
\description{
Add an extra column to an existing \code{\link{big.data.frame}}
}

