% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{big.data.frame-class}
\alias{big.data.frame-class}
\title{S4 class big.data.frame is essentially a list of big vectors
provided by packages bigmemory and big.char.}
\description{
S4 class big.data.frame is essentially a list of big vectors
provided by packages bigmemory and big.char.
}

