% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ncol,big.data.frame-method}
\alias{$,big.data.frame-method}
\alias{$<-,big.data.frame-method}
\alias{[,big.data.frame,ANY,ANY,missing-method}
\alias{[,big.data.frame,ANY,missing,missing-method}
\alias{[,big.data.frame,missing,ANY,logical-method}
\alias{[,big.data.frame,missing,ANY,missing-method}
\alias{[,big.data.frame,missing,missing,missing-method}
\alias{[<-,big.data.frame,ANY,missing-method}
\alias{[<-,big.data.frame,missing,ANY-method}
\alias{dim,big.data.frame-method}
\alias{head.big.data.frame}
\alias{is.big.data.frame,big.data.frame-method}
\alias{length,big.data.frame-method}
\alias{names,big.data.frame-method}
\alias{names<-,big.data.frame,character-method}
\alias{ncol,big.data.frame-method}
\alias{nrow,big.data.frame-method}
\alias{str,big.data.frame-method}
\alias{summary,big.data.frame-method}
\alias{tail.big.data.frame}
\title{ncol functionality for a big.data.frame}
\usage{
\S4method{ncol}{big.data.frame}(x)

\S4method{nrow}{big.data.frame}(x)

\S4method{dim}{big.data.frame}(x)

\S4method{length}{big.data.frame}(x)

\method{head}{big.data.frame}(x, n = 6)

\method{tail}{big.data.frame}(x, n = 6)

\S4method{summary}{big.data.frame}(object, ...)

\S4method{str}{big.data.frame}(object, ...)

\S4method{names}{big.data.frame}(x)

\S4method{names}{big.data.frame,character}(x) <- value

\S4method{is.big.data.frame}{big.data.frame}(x)

\S4method{[}{big.data.frame,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{big.data.frame,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{big.data.frame,ANY,missing}(x, i, j, ...) <- value

\S4method{[}{big.data.frame,missing,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{big.data.frame,missing,ANY}(x, i, j, ...) <- value

\S4method{[}{big.data.frame,missing,ANY,logical}(x, i, j, ..., drop = TRUE)

\S4method{[}{big.data.frame,missing,missing,missing}(x, i, j, ...,
  drop = TRUE)

\S4method{$}{big.data.frame}(x, name)

\S4method{$}{big.data.frame}(x, name) <- value
}
\arguments{
\item{x}{a big.data.frame}

\item{x}{a big.data.frame}

\item{x}{a big.data.frame}
}
\description{
ncol functionality for a big.data.frame

nrow functionality for a big.data.frame

dim functionality for a big.data.frame

length functionality for a big.data.frame

length functionality for a big.data.frame

length functionality for a big.data.frame

print summary of the big data frame

print structure of the big data frame

names functionality for a big.data.frame

set the names of a big.data.frame

Do we have a big.data.frame?
}
\author{
Miranda Sinnott-Armstrong

Miranda Sinnott-Armstrong

Rose Brewin

Rose Brewin

Rose Brewin

Rose Brewin

Rose Brewin

Miranda Sinnott-Armstrong

Miranda Sinnott-Armstrong

Miranda Sinnott-Armstrong

Miranda Sinnott-Armstrong
}

