% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{big.data.frame-package}
\alias{big.data.frame-package}
\title{Provides very large data.frame-like objects}
\description{
This package extends bigmemory's big.matrix objects along with
new big.char objects to support larger-than-RAM data.frame-like
objects in R.
}
\author{
Jay Emerson
}

