% Generated by roxygen2 (4.0.2): do not edit by hand
\name{big.data.frame}
\alias{big.data.frame}
\title{Create a big.data.frame}
\usage{
big.data.frame(nrow, classes, location = NULL, names = NULL,
  maxchar = NULL, init = NULL)
}
\arguments{
\item{nrow}{the number of rows}

\item{classes}{a vector of values from 'double', 'character', ...}

\item{location}{folder to contain the object backingfiles, or NULL
if the object will be in-RAM.}

\item{names}{a vector of names for the columns (variables)}

\item{maxchar}{a vector with NA for numeric columns, but integers
specifying the maximum number of characters for string columns (see
\code{\link[big.char]{big.char}} for more information).}

\item{init}{a vector of values for initialization; note that for
large objects this will slow down the creation substantially.}
}
\value{
Returns a \code{big.data.frame} object
}
\description{
Create a \code{big.data.frame}
}
\details{
This is the full set of details for documentation.
}
\author{
Jay Emerson
}

