% Generated by roxygen2 (4.0.2): do not edit by hand
\name{big.drop.cols}
\alias{big.drop.cols}
\title{Remove columns from an existing \code{\link{big.data.frame}}}
\usage{
big.drop.cols(x, index, location = NULL)
}
\arguments{
\item{x}{a big.data.frame object}

\item{index}{a vector of indices or names of the columns to be removed}
}
\value{
a new \code{\link{big.data.frame}} object, with fewer columns
}
\description{
Remove columns from an existing \code{\link{big.data.frame}}
}
\author{
Rose Brewin
}

