% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.big.data.frame}
\alias{is.big.data.frame}
\title{Generic function is.big.data.frame()}
\usage{
is.big.data.frame(x)
}
\arguments{
\item{x}{a \code{\link{big.data.frame}} combination of big.matrix
and big.char objects}
}
\description{
Do we have a \code{\link{big.data.frame}}?
}
\details{
No further detail is needed.
}
\author{
Rose Brewin
}

