% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ncol,big.data.frame-method}
\alias{dim,big.data.frame-method}
\alias{length,big.data.frame-method}
\alias{ncol,big.data.frame-method}
\alias{nrow,big.data.frame-method}
\title{ncol functionality for a big.data.frame}
\usage{
\S4method{ncol}{big.data.frame}(x)

\S4method{nrow}{big.data.frame}(x)

\S4method{dim}{big.data.frame}(x)

\S4method{length}{big.data.frame}(x)
}
\arguments{
\item{x}{a big.data.frame}
}
\description{
ncol functionality for a big.data.frame

nrow functionality for a big.data.frame

dim functionality for a big.data.frame

length functionality for a big.data.frame
}

